backend/requirements.txt
Django==5.1.6
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-allauth==0.61.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1 